authorization and authentication done

+++ controllers/api/v1/posts_api.js
@@ -1,5 +1,6 @@
 const Post = require('../../../models/post');
 const Comment = require('../../../models/comment');
+
 module.exports.index = async function (req, res) {
 
     let posts = await Post.find({})
@@ -22,25 +23,26 @@ module.exports.destroy = async function (req, res) {
     try {
         let post = await Post.findById(req.params.id);
 
-        // if (post.user == req.user.id) {
+        if (post.user == req.user.id) {
             post.remove();
             // post.save();
 
             await Comment.deleteMany({
                 post: req.params.id
             });
-            // req.flash('success','post deleted Successfully');
-            return res.json(200,{
-                message:"post and associated comment deleted successfully"
+            
+            return res.json(200, {
+                message: "post and associated comment deleted successfully"
             });
-        // } 
-        // else {
-        //     req.flash('error','post deleted unSuccessfully');
-        //     return res.redirect('back');
-        // }
+        } else {
+            return res.json(401, {
+                message: "you can not delete this post"
+            });
+        }
     } catch (err) {
-        return res.json(500,{
-            message:"Internal server error"
+        console.log("**********",err);
+        return res.json(500, {
+            message: "Internal server error"
         });
     }
 }
